Moved DOM Utils to independent file
